Попробовал убрать из Program работу с контейнером, оставил только конфигурирование. Убрал в TagsCloudApp.Run() , но не пойму как зарезолвить объекты
