+++ TagsCloudVisualization/TagsCloudVisualization/Interfaces/IWordDataProvider.cs
@@ -4,6 +4,6 @@ namespace TagsCloudVisualization.Interfaces
 {
     public interface IWordDataProvider
     {
-        List<CloudWordData> GetData(ICloudLayouter cloud, string parametersFilePath);
+        List<CloudWordData> GetData(ICloudLayouter cloud, List<string> words);
     }
 }
