+++ TagsCloudVisualization/TagsCloudVisualization/WordsExtractor.cs
@@ -2,7 +2,6 @@
 using System.IO;
 using System.Linq;
 using System.Text;
-using System.Windows.Forms.VisualStyles;
 using TagsCloudVisualization.Interfaces;
 
 namespace TagsCloudVisualization
